+++ core/config.py
@@ -52,6 +52,7 @@ class ConfigManager:
         "close_emoji": "\N{LOCK}",
         "use_user_id_channel_name": False,
         "use_timestamp_channel_name": False,
+        "use_nickname_channel_name": False,
         "recipient_thread_close": False,
         "thread_show_roles": True,
         "thread_show_account_age": True,
@@ -77,7 +78,7 @@ class ConfigManager:
         "thread_move_notify_mods": False,
         "thread_move_response": "This thread has been moved.",
         "cooldown_thread_title": "Message not sent!",
-        "cooldown_thread_response": "You must wait for {delta} before you can contact me again.",
+        "cooldown_thread_response": "Your cooldown ends {delta}. Try contacting me then.",
         "disabled_new_thread_title": "Not Delivered",
         "disabled_new_thread_response": "We are not accepting new threads.",
         "disabled_new_thread_footer": "Please try again later...",
@@ -91,6 +92,9 @@ class ConfigManager:
         "silent_alert_on_mention": False,
         "show_timestamp": True,
         "anonymous_snippets": False,
+        "plain_snippets": False,
+        "require_close_reason": False,
+        "show_log_url_button": False,
         # group conversations
         "private_added_to_group_title": "New Thread (Group)",
         "private_added_to_group_response": "{moderator.name} has added you to a Modmail thread.",
@@ -123,6 +127,7 @@ class ConfigManager:
         "confirm_thread_creation_deny": "\N{NO ENTRY SIGN}",
         # regex
         "use_regex_autotrigger": False,
+        "use_hoisted_top_role": True,
     }
 
     private_keys = {
@@ -181,11 +186,13 @@ class ConfigManager:
     booleans = {
         "use_user_id_channel_name",
         "use_timestamp_channel_name",
+        "use_nickname_channel_name",
         "user_typing",
         "mod_typing",
         "reply_without_command",
         "anon_reply_without_command",
         "plain_reply_without_command",
+        "show_log_url_button",
         "recipient_thread_close",
         "thread_auto_close_silently",
         "thread_move_notify",
@@ -205,10 +212,13 @@ class ConfigManager:
         "update_notifications",
         "thread_contact_silently",
         "anonymous_snippets",
+        "plain_snippets",
+        "require_close_reason",
         "recipient_thread_close",
         "thread_show_roles",
         "thread_show_account_age",
         "thread_show_join_age",
+        "use_hoisted_top_role",
     }
 
     enums = {